add consistency constraints for correct limb decomposition to Hash Table
Also:
- fix bug: correctly compute inverses
- fix bug: correctly set inverses when padding Hash Table
- move construction of Hash Table rows from `vm.rs` to `hash_table.rs`
