+++ triton-vm/src/table/hash_table.rs
@@ -1,13 +1,14 @@
 use itertools::Itertools;
 use ndarray::s;
+use ndarray::Array2;
 use ndarray::ArrayView1;
 use ndarray::ArrayView2;
 use ndarray::ArrayViewMut2;
 use num_traits::Zero;
 use strum::EnumCount;
-use triton_opcodes::instruction::Instruction;
 use twenty_first::shared_math::b_field_element::BFieldElement;
 use twenty_first::shared_math::b_field_element::BFIELD_ONE;
+use twenty_first::shared_math::b_field_element::BFIELD_ZERO;
 use twenty_first::shared_math::tip5::DIGEST_LENGTH;
 use twenty_first::shared_math::tip5::MDS_MATRIX_FIRST_COLUMN;
 use twenty_first::shared_math::tip5::NUM_ROUNDS;
@@ -15,8 +16,12 @@ use twenty_first::shared_math::tip5::NUM_SPLIT_AND_LOOKUP;
 use twenty_first::shared_math::tip5::RATE;
 use twenty_first::shared_math::tip5::ROUND_CONSTANTS;
 use twenty_first::shared_math::tip5::STATE_SIZE;
+use twenty_first::shared_math::traits::Inverse;
 use twenty_first::shared_math::x_field_element::XFieldElement;
 
+use triton_opcodes::instruction::Instruction;
+
+use crate::table::cascade_table::CascadeTable;
 use crate::table::challenges::ChallengeId::*;
 use crate::table::challenges::Challenges;
 use crate::table::constraint_circuit::ConstraintCircuit;
@@ -258,6 +263,77 @@ impl ExtHashTable {
         let if_ci_is_absorb_init_and_round_no_is_0_then_state_15_is_0 =
             if_ci_is_absorb_init_and_round_no_is_0_then_ * state15;
 
+        // consistency of the inverse of the highest 2 limbs minus 2^32 - 1
+        let one = constant(1);
+        let two_pow_16 = constant(1 << 16);
+        let two_pow_32 = constant(1 << 32);
+        let state_0_hi_limbs_minus_2_pow_32 = two_pow_32.clone()
+            - one.clone()
+            - base_row(State0HighestLkIn) * two_pow_16.clone()
+            - base_row(State0MidHighLkIn);
+        let state_1_hi_limbs_minus_2_pow_32 = two_pow_32.clone()
+            - one.clone()
+            - base_row(State1HighestLkIn) * two_pow_16.clone()
+            - base_row(State1MidHighLkIn);
+        let state_2_hi_limbs_minus_2_pow_32 = two_pow_32.clone()
+            - one.clone()
+            - base_row(State2HighestLkIn) * two_pow_16.clone()
+            - base_row(State2MidHighLkIn);
+        let state_3_hi_limbs_minus_2_pow_32 = two_pow_32.clone()
+            - one.clone()
+            - base_row(State3HighestLkIn) * two_pow_16.clone()
+            - base_row(State3MidHighLkIn);
+
+        let state_0_hi_limbs_inv = base_row(State0Inv);
+        let state_1_hi_limbs_inv = base_row(State1Inv);
+        let state_2_hi_limbs_inv = base_row(State2Inv);
+        let state_3_hi_limbs_inv = base_row(State3Inv);
+
+        let state_0_hi_limbs_are_not_all_1s =
+            state_0_hi_limbs_minus_2_pow_32.clone() * state_0_hi_limbs_inv.clone() - one.clone();
+        let state_1_hi_limbs_are_not_all_1s =
+            state_1_hi_limbs_minus_2_pow_32.clone() * state_1_hi_limbs_inv.clone() - one.clone();
+        let state_2_hi_limbs_are_not_all_1s =
+            state_2_hi_limbs_minus_2_pow_32.clone() * state_2_hi_limbs_inv.clone() - one.clone();
+        let state_3_hi_limbs_are_not_all_1s =
+            state_3_hi_limbs_minus_2_pow_32.clone() * state_3_hi_limbs_inv.clone() - one;
+
+        let state_0_hi_limbs_inv_is_inv_or_is_zero =
+            state_0_hi_limbs_are_not_all_1s.clone() * state_0_hi_limbs_inv;
+        let state_1_hi_limbs_inv_is_inv_or_is_zero =
+            state_1_hi_limbs_are_not_all_1s.clone() * state_1_hi_limbs_inv;
+        let state_2_hi_limbs_inv_is_inv_or_is_zero =
+            state_2_hi_limbs_are_not_all_1s.clone() * state_2_hi_limbs_inv;
+        let state_3_hi_limbs_inv_is_inv_or_is_zero =
+            state_3_hi_limbs_are_not_all_1s.clone() * state_3_hi_limbs_inv;
+
+        let state_0_hi_limbs_inv_is_inv_or_state_0_hi_limbs_is_zero =
+            state_0_hi_limbs_are_not_all_1s.clone() * state_0_hi_limbs_minus_2_pow_32;
+        let state_1_hi_limbs_inv_is_inv_or_state_1_hi_limbs_is_zero =
+            state_1_hi_limbs_are_not_all_1s.clone() * state_1_hi_limbs_minus_2_pow_32;
+        let state_2_hi_limbs_inv_is_inv_or_state_2_hi_limbs_is_zero =
+            state_2_hi_limbs_are_not_all_1s.clone() * state_2_hi_limbs_minus_2_pow_32;
+        let state_3_hi_limbs_inv_is_inv_or_state_3_hi_limbs_is_zero =
+            state_3_hi_limbs_are_not_all_1s.clone() * state_3_hi_limbs_minus_2_pow_32;
+
+        // consistent decomposition into limbs
+        let state_0_lo_limbs =
+            base_row(State0MidLowLkIn) * two_pow_16.clone() + base_row(State0LowestLkIn);
+        let state_1_lo_limbs =
+            base_row(State1MidLowLkIn) * two_pow_16.clone() + base_row(State1LowestLkIn);
+        let state_2_lo_limbs =
+            base_row(State2MidLowLkIn) * two_pow_16.clone() + base_row(State2LowestLkIn);
+        let state_3_lo_limbs = base_row(State3MidLowLkIn) * two_pow_16 + base_row(State3LowestLkIn);
+
+        let if_state_0_hi_limbs_are_all_1_then_state_0_lo_limbs_are_all_0 =
+            state_0_hi_limbs_are_not_all_1s * state_0_lo_limbs;
+        let if_state_1_hi_limbs_are_all_1_then_state_1_lo_limbs_are_all_0 =
+            state_1_hi_limbs_are_not_all_1s * state_1_lo_limbs;
+        let if_state_2_hi_limbs_are_all_1_then_state_2_lo_limbs_are_all_0 =
+            state_2_hi_limbs_are_not_all_1s * state_2_lo_limbs;
+        let if_state_3_hi_limbs_are_all_1_then_state_3_lo_limbs_are_all_0 =
+            state_3_hi_limbs_are_not_all_1s * state_3_lo_limbs;
+
         let mut constraints = vec![
             if_padding_row_then_ci_is_hash,
             if_ci_is_hash_and_round_no_is_0_then_state_10_is_1,
@@ -272,6 +348,18 @@ impl ExtHashTable {
             if_ci_is_absorb_init_and_round_no_is_0_then_state_13_is_0,
             if_ci_is_absorb_init_and_round_no_is_0_then_state_14_is_0,
             if_ci_is_absorb_init_and_round_no_is_0_then_state_15_is_0,
+            state_0_hi_limbs_inv_is_inv_or_is_zero,
+            state_1_hi_limbs_inv_is_inv_or_is_zero,
+            state_2_hi_limbs_inv_is_inv_or_is_zero,
+            state_3_hi_limbs_inv_is_inv_or_is_zero,
+            state_0_hi_limbs_inv_is_inv_or_state_0_hi_limbs_is_zero,
+            state_1_hi_limbs_inv_is_inv_or_state_1_hi_limbs_is_zero,
+            state_2_hi_limbs_inv_is_inv_or_state_2_hi_limbs_is_zero,
+            state_3_hi_limbs_inv_is_inv_or_state_3_hi_limbs_is_zero,
+            if_state_0_hi_limbs_are_all_1_then_state_0_lo_limbs_are_all_0,
+            if_state_1_hi_limbs_are_all_1_then_state_1_lo_limbs_are_all_0,
+            if_state_2_hi_limbs_are_all_1_then_state_2_lo_limbs_are_all_0,
+            if_state_3_hi_limbs_are_all_1_then_state_3_lo_limbs_are_all_0,
         ];
 
         for round_constant_column_idx in 0..NUM_ROUND_CONSTANTS {
@@ -774,6 +862,151 @@ impl HashTable {
         BFieldElement::new(mds_matrix_entry as u64)
     }
 
+    /// The round constants for round `round_number` if it is a valid round number in the Tip5
+    /// permutation, and the zero vector otherwise.
+    pub fn tip5_round_constants_by_round_number(
+        round_number: usize,
+    ) -> [BFieldElement; NUM_ROUND_CONSTANTS] {
+        match round_number {
+            i if i < NUM_ROUNDS => ROUND_CONSTANTS
+                [NUM_ROUND_CONSTANTS * i..NUM_ROUND_CONSTANTS * (i + 1)]
+                .try_into()
+                .unwrap(),
+            _ => [BFIELD_ZERO; NUM_ROUND_CONSTANTS],
+        }
+    }
+
+    /// Given a trace of the Tip5 permutation, construct a trace corresponding to the columns of
+    /// the Hash Table. This includes
+    ///
+    /// - adding the round number
+    /// - adding the round constants,
+    /// - decomposing the first [`NUM_SPLIT_AND_LOOKUP`] (== 4) state elements into their
+    ///     constituent limbs,
+    /// - setting the inverse-or-zero for proving correct limb decomposition, and
+    /// - adding the looked-up value for each limb.
+    ///
+    /// The current instruction is not set.
+    pub fn convert_to_hash_table_rows(
+        hash_permutation_trace: [[BFieldElement; STATE_SIZE]; NUM_ROUNDS + 1],
+    ) -> Array2<BFieldElement> {
+        let mut hash_trace_addendum = Array2::zeros([NUM_ROUNDS + 1, BASE_WIDTH]);
+        for (round_number, mut row) in hash_trace_addendum.rows_mut().into_iter().enumerate() {
+            let trace_row = hash_permutation_trace[round_number];
+            row[RoundNumber.base_table_index()] = BFieldElement::from(round_number as u64);
+
+            let st_0_raw_limbs = trace_row[0].raw_u16s();
+            let st_0_look_in_split =
+                st_0_raw_limbs.map(|limb| BFieldElement::from_raw_u64(limb as u64));
+            row[State0LowestLkIn.base_table_index()] = st_0_look_in_split[0];
+            row[State0MidLowLkIn.base_table_index()] = st_0_look_in_split[1];
+            row[State0MidHighLkIn.base_table_index()] = st_0_look_in_split[2];
+            row[State0HighestLkIn.base_table_index()] = st_0_look_in_split[3];
+
+            let st_0_look_out_split = st_0_raw_limbs.map(CascadeTable::lookup_16_bit_limb);
+            row[State0LowestLkOut.base_table_index()] = st_0_look_out_split[0];
+            row[State0MidLowLkOut.base_table_index()] = st_0_look_out_split[1];
+            row[State0MidHighLkOut.base_table_index()] = st_0_look_out_split[2];
+            row[State0HighestLkOut.base_table_index()] = st_0_look_out_split[3];
+
+            let st_1_raw_limbs = trace_row[1].raw_u16s();
+            let st_1_look_in_split =
+                st_1_raw_limbs.map(|limb| BFieldElement::from_raw_u64(limb as u64));
+            row[State1LowestLkIn.base_table_index()] = st_1_look_in_split[0];
+            row[State1MidLowLkIn.base_table_index()] = st_1_look_in_split[1];
+            row[State1MidHighLkIn.base_table_index()] = st_1_look_in_split[2];
+            row[State1HighestLkIn.base_table_index()] = st_1_look_in_split[3];
+
+            let st_1_look_out_split = st_1_raw_limbs.map(CascadeTable::lookup_16_bit_limb);
+            row[State1LowestLkOut.base_table_index()] = st_1_look_out_split[0];
+            row[State1MidLowLkOut.base_table_index()] = st_1_look_out_split[1];
+            row[State1MidHighLkOut.base_table_index()] = st_1_look_out_split[2];
+            row[State1HighestLkOut.base_table_index()] = st_1_look_out_split[3];
+
+            let st_2_raw_limbs = trace_row[2].raw_u16s();
+            let st_2_look_in_split =
+                st_2_raw_limbs.map(|limb| BFieldElement::from_raw_u64(limb as u64));
+            row[State2LowestLkIn.base_table_index()] = st_2_look_in_split[0];
+            row[State2MidLowLkIn.base_table_index()] = st_2_look_in_split[1];
+            row[State2MidHighLkIn.base_table_index()] = st_2_look_in_split[2];
+            row[State2HighestLkIn.base_table_index()] = st_2_look_in_split[3];
+
+            let st_2_look_out_split = st_2_raw_limbs.map(CascadeTable::lookup_16_bit_limb);
+            row[State2LowestLkOut.base_table_index()] = st_2_look_out_split[0];
+            row[State2MidLowLkOut.base_table_index()] = st_2_look_out_split[1];
+            row[State2MidHighLkOut.base_table_index()] = st_2_look_out_split[2];
+            row[State2HighestLkOut.base_table_index()] = st_2_look_out_split[3];
+
+            let st_3_raw_limbs = trace_row[3].raw_u16s();
+            let st_3_look_in_split =
+                st_3_raw_limbs.map(|limb| BFieldElement::from_raw_u64(limb as u64));
+            row[State3LowestLkIn.base_table_index()] = st_3_look_in_split[0];
+            row[State3MidLowLkIn.base_table_index()] = st_3_look_in_split[1];
+            row[State3MidHighLkIn.base_table_index()] = st_3_look_in_split[2];
+            row[State3HighestLkIn.base_table_index()] = st_3_look_in_split[3];
+
+            let st_3_look_out_split = st_3_raw_limbs.map(CascadeTable::lookup_16_bit_limb);
+            row[State3LowestLkOut.base_table_index()] = st_3_look_out_split[0];
+            row[State3MidLowLkOut.base_table_index()] = st_3_look_out_split[1];
+            row[State3MidHighLkOut.base_table_index()] = st_3_look_out_split[2];
+            row[State3HighestLkOut.base_table_index()] = st_3_look_out_split[3];
+
+            row[State4.base_table_index()] = trace_row[4];
+            row[State5.base_table_index()] = trace_row[5];
+            row[State6.base_table_index()] = trace_row[6];
+            row[State7.base_table_index()] = trace_row[7];
+            row[State8.base_table_index()] = trace_row[8];
+            row[State9.base_table_index()] = trace_row[9];
+            row[State10.base_table_index()] = trace_row[10];
+            row[State11.base_table_index()] = trace_row[11];
+            row[State12.base_table_index()] = trace_row[12];
+            row[State13.base_table_index()] = trace_row[13];
+            row[State14.base_table_index()] = trace_row[14];
+            row[State15.base_table_index()] = trace_row[15];
+
+            row[State0Inv.base_table_index()] =
+                Self::inverse_or_zero_of_highest_2_limbs(trace_row[0]);
+            row[State1Inv.base_table_index()] =
+                Self::inverse_or_zero_of_highest_2_limbs(trace_row[1]);
+            row[State2Inv.base_table_index()] =
+                Self::inverse_or_zero_of_highest_2_limbs(trace_row[2]);
+            row[State3Inv.base_table_index()] =
+                Self::inverse_or_zero_of_highest_2_limbs(trace_row[3]);
+
+            let round_constants = Self::tip5_round_constants_by_round_number(round_number);
+            row[Constant0.base_table_index()] = round_constants[0];
+            row[Constant1.base_table_index()] = round_constants[1];
+            row[Constant2.base_table_index()] = round_constants[2];
+            row[Constant3.base_table_index()] = round_constants[3];
+            row[Constant4.base_table_index()] = round_constants[4];
+            row[Constant5.base_table_index()] = round_constants[5];
+            row[Constant6.base_table_index()] = round_constants[6];
+            row[Constant7.base_table_index()] = round_constants[7];
+            row[Constant8.base_table_index()] = round_constants[8];
+            row[Constant9.base_table_index()] = round_constants[9];
+            row[Constant10.base_table_index()] = round_constants[10];
+            row[Constant11.base_table_index()] = round_constants[11];
+            row[Constant12.base_table_index()] = round_constants[12];
+            row[Constant13.base_table_index()] = round_constants[13];
+            row[Constant14.base_table_index()] = round_constants[14];
+            row[Constant15.base_table_index()] = round_constants[15];
+        }
+        hash_trace_addendum
+    }
+
+    /// The inverse-or-zero of (2^32 - 1 - 2^16·`highest` - `mid_high`) where `highest`
+    /// is the most significant limb of the given `state_element`, and `mid_high` the second-most
+    /// significant limb.
+    fn inverse_or_zero_of_highest_2_limbs(state_element: BFieldElement) -> BFieldElement {
+        let limbs = state_element.raw_u16s();
+        let highest = limbs[3];
+        let mid_high = limbs[2];
+        let high_limbs = BFieldElement::from_raw_u16s(&[mid_high, highest, 0, 0]);
+        let two_pow_32_minus_1 = BFieldElement::new((1 << 32) - 1);
+        let to_invert = two_pow_32_minus_1 - high_limbs;
+        to_invert.inverse_or_zero()
+    }
+
     pub fn fill_trace(
         hash_table: &mut ArrayViewMut2<BFieldElement>,
         aet: &AlgebraicExecutionTrace,
@@ -796,6 +1029,21 @@ impl HashTable {
         hash_table
             .slice_mut(s![hash_table_length.., RoundNumber.base_table_index()])
             .fill(-BFIELD_ONE);
+
+        let inverse_of_high_limbs_plus_1_minus_2_pow_32 =
+            Self::inverse_or_zero_of_highest_2_limbs(BFIELD_ZERO);
+        hash_table
+            .slice_mut(s![hash_table_length.., State0Inv.base_table_index()])
+            .fill(inverse_of_high_limbs_plus_1_minus_2_pow_32);
+        hash_table
+            .slice_mut(s![hash_table_length.., State1Inv.base_table_index()])
+            .fill(inverse_of_high_limbs_plus_1_minus_2_pow_32);
+        hash_table
+            .slice_mut(s![hash_table_length.., State2Inv.base_table_index()])
+            .fill(inverse_of_high_limbs_plus_1_minus_2_pow_32);
+        hash_table
+            .slice_mut(s![hash_table_length.., State3Inv.base_table_index()])
+            .fill(inverse_of_high_limbs_plus_1_minus_2_pow_32);
     }
 
     pub fn extend(
