+++ triton-vm/src/vm.rs
@@ -3,7 +3,6 @@ use std::convert::TryInto;
 use std::fmt::Display;
 
 use anyhow::Result;
-use itertools::Itertools;
 use ndarray::s;
 use ndarray::Array1;
 use ndarray::Array2;
@@ -11,7 +10,6 @@ use ndarray::Axis;
 use num_traits::One;
 use num_traits::Zero;
 use twenty_first::shared_math::b_field_element::BFieldElement;
-use twenty_first::shared_math::b_field_element::BFIELD_ZERO;
 use twenty_first::shared_math::other::log_2_floor;
 use twenty_first::shared_math::tip5;
 use twenty_first::shared_math::tip5::Tip5;
@@ -33,9 +31,8 @@ use crate::error::vm_fail;
 use crate::error::InstructionError::InstructionPointerOverflow;
 use crate::error::InstructionError::*;
 use crate::op_stack::OpStack;
-use crate::table::cascade_table::CascadeTable;
 use crate::table::hash_table;
-use crate::table::hash_table::NUM_ROUND_CONSTANTS;
+use crate::table::hash_table::HashTable;
 use crate::table::processor_table;
 use crate::table::processor_table::ProcessorTraceRow;
 use crate::table::table_column::HashBaseTableColumn::*;
@@ -967,7 +964,7 @@ impl AlgebraicExecutionTrace {
         hash_permutation_trace: [[BFieldElement; tip5::STATE_SIZE]; tip5::NUM_ROUNDS + 1],
     ) {
         self.increase_lookup_multiplicities(hash_permutation_trace);
-        let mut hash_trace_addendum = Self::convert_to_hash_table_rows(hash_permutation_trace);
+        let mut hash_trace_addendum = HashTable::convert_to_hash_table_rows(hash_permutation_trace);
         hash_trace_addendum
             .slice_mut(s![.., CI.base_table_index()])
             .fill(Instruction::Hash.opcode_b());
@@ -986,7 +983,8 @@ impl AlgebraicExecutionTrace {
             Instruction::AbsorbInit | Instruction::Absorb | Instruction::Squeeze
         ));
         self.increase_lookup_multiplicities(hash_permutation_trace);
-        let mut sponge_trace_addendum = Self::convert_to_hash_table_rows(hash_permutation_trace);
+        let mut sponge_trace_addendum =
+            HashTable::convert_to_hash_table_rows(hash_permutation_trace);
         sponge_trace_addendum
             .slice_mut(s![.., CI.base_table_index()])
             .fill(instruction.opcode_b());
@@ -1017,149 +1015,6 @@ impl AlgebraicExecutionTrace {
                 }
             }
         }
-    }
-
-    /// Given a trace of the Tip5 permutation, construct a trace corresponding to the columns of
-    /// the Hash Table. This includes
-    ///
-    /// - adding the round number
-    /// - adding the round constants,
-    /// - decomposing the first [`tip5::NUM_SPLIT_AND_LOOKUP`] (== 4) state elements into their
-    ///     constituent limbs,
-    /// - setting the inverse-or-zero for proving correct limb decomposition, and
-    /// - adding the looked-up value for each limb.
-    ///
-    /// The current instruction is not set.
-    fn convert_to_hash_table_rows(
-        hash_permutation_trace: [[BFieldElement; tip5::STATE_SIZE]; tip5::NUM_ROUNDS + 1],
-    ) -> Array2<BFieldElement> {
-        let mut hash_trace_addendum = Array2::zeros([tip5::NUM_ROUNDS + 1, hash_table::BASE_WIDTH]);
-        for (round_number, mut row) in hash_trace_addendum.rows_mut().into_iter().enumerate() {
-            let trace_row = hash_permutation_trace[round_number];
-            row[RoundNumber.base_table_index()] = BFieldElement::from(round_number as u64);
-
-            let st_0_raw_limbs = trace_row[0].raw_u16s();
-            let st_0_look_in_split =
-                st_0_raw_limbs.map(|limb| BFieldElement::from_raw_u64(limb as u64));
-            row[State0LowestLkIn.base_table_index()] = st_0_look_in_split[0];
-            row[State0MidLowLkIn.base_table_index()] = st_0_look_in_split[1];
-            row[State0MidHighLkIn.base_table_index()] = st_0_look_in_split[2];
-            row[State0HighestLkIn.base_table_index()] = st_0_look_in_split[3];
-
-            let st_0_look_out_split = st_0_raw_limbs.map(CascadeTable::lookup_16_bit_limb);
-            row[State0LowestLkOut.base_table_index()] = st_0_look_out_split[0];
-            row[State0MidLowLkOut.base_table_index()] = st_0_look_out_split[1];
-            row[State0MidHighLkOut.base_table_index()] = st_0_look_out_split[2];
-            row[State0HighestLkOut.base_table_index()] = st_0_look_out_split[3];
-
-            let st_1_raw_limbs = trace_row[1].raw_u16s();
-            let st_1_look_in_split =
-                st_1_raw_limbs.map(|limb| BFieldElement::from_raw_u64(limb as u64));
-            row[State1LowestLkIn.base_table_index()] = st_1_look_in_split[0];
-            row[State1MidLowLkIn.base_table_index()] = st_1_look_in_split[1];
-            row[State1MidHighLkIn.base_table_index()] = st_1_look_in_split[2];
-            row[State1HighestLkIn.base_table_index()] = st_1_look_in_split[3];
-
-            let st_1_look_out_split = st_1_raw_limbs.map(CascadeTable::lookup_16_bit_limb);
-            row[State1LowestLkOut.base_table_index()] = st_1_look_out_split[0];
-            row[State1MidLowLkOut.base_table_index()] = st_1_look_out_split[1];
-            row[State1MidHighLkOut.base_table_index()] = st_1_look_out_split[2];
-            row[State1HighestLkOut.base_table_index()] = st_1_look_out_split[3];
-
-            let st_2_raw_limbs = trace_row[2].raw_u16s();
-            let st_2_look_in_split =
-                st_2_raw_limbs.map(|limb| BFieldElement::from_raw_u64(limb as u64));
-            row[State2LowestLkIn.base_table_index()] = st_2_look_in_split[0];
-            row[State2MidLowLkIn.base_table_index()] = st_2_look_in_split[1];
-            row[State2MidHighLkIn.base_table_index()] = st_2_look_in_split[2];
-            row[State2HighestLkIn.base_table_index()] = st_2_look_in_split[3];
-
-            let st_2_look_out_split = st_2_raw_limbs.map(CascadeTable::lookup_16_bit_limb);
-            row[State2LowestLkOut.base_table_index()] = st_2_look_out_split[0];
-            row[State2MidLowLkOut.base_table_index()] = st_2_look_out_split[1];
-            row[State2MidHighLkOut.base_table_index()] = st_2_look_out_split[2];
-            row[State2HighestLkOut.base_table_index()] = st_2_look_out_split[3];
-
-            let st_3_raw_limbs = trace_row[3].raw_u16s();
-            let st_3_look_in_split =
-                st_3_raw_limbs.map(|limb| BFieldElement::from_raw_u64(limb as u64));
-            row[State3LowestLkIn.base_table_index()] = st_3_look_in_split[0];
-            row[State3MidLowLkIn.base_table_index()] = st_3_look_in_split[1];
-            row[State3MidHighLkIn.base_table_index()] = st_3_look_in_split[2];
-            row[State3HighestLkIn.base_table_index()] = st_3_look_in_split[3];
-
-            let st_3_look_out_split = st_3_raw_limbs.map(CascadeTable::lookup_16_bit_limb);
-            row[State3LowestLkOut.base_table_index()] = st_3_look_out_split[0];
-            row[State3MidLowLkOut.base_table_index()] = st_3_look_out_split[1];
-            row[State3MidHighLkOut.base_table_index()] = st_3_look_out_split[2];
-            row[State3HighestLkOut.base_table_index()] = st_3_look_out_split[3];
-
-            row[State4.base_table_index()] = trace_row[4];
-            row[State5.base_table_index()] = trace_row[5];
-            row[State6.base_table_index()] = trace_row[6];
-            row[State7.base_table_index()] = trace_row[7];
-            row[State8.base_table_index()] = trace_row[8];
-            row[State9.base_table_index()] = trace_row[9];
-            row[State10.base_table_index()] = trace_row[10];
-            row[State11.base_table_index()] = trace_row[11];
-            row[State12.base_table_index()] = trace_row[12];
-            row[State13.base_table_index()] = trace_row[13];
-            row[State14.base_table_index()] = trace_row[14];
-            row[State15.base_table_index()] = trace_row[15];
-
-            row[State0Inv.base_table_index()] =
-                Self::inverse_or_zero_of_highest_2_limbs(trace_row[0]);
-            row[State1Inv.base_table_index()] =
-                Self::inverse_or_zero_of_highest_2_limbs(trace_row[1]);
-            row[State2Inv.base_table_index()] =
-                Self::inverse_or_zero_of_highest_2_limbs(trace_row[2]);
-            row[State3Inv.base_table_index()] =
-                Self::inverse_or_zero_of_highest_2_limbs(trace_row[3]);
-
-            let round_constants = Self::tip5_round_constants_by_round_number(round_number);
-            row[Constant0.base_table_index()] = round_constants[0];
-            row[Constant1.base_table_index()] = round_constants[1];
-            row[Constant2.base_table_index()] = round_constants[2];
-            row[Constant3.base_table_index()] = round_constants[3];
-            row[Constant4.base_table_index()] = round_constants[4];
-            row[Constant5.base_table_index()] = round_constants[5];
-            row[Constant6.base_table_index()] = round_constants[6];
-            row[Constant7.base_table_index()] = round_constants[7];
-            row[Constant8.base_table_index()] = round_constants[8];
-            row[Constant9.base_table_index()] = round_constants[9];
-            row[Constant10.base_table_index()] = round_constants[10];
-            row[Constant11.base_table_index()] = round_constants[11];
-            row[Constant12.base_table_index()] = round_constants[12];
-            row[Constant13.base_table_index()] = round_constants[13];
-            row[Constant14.base_table_index()] = round_constants[14];
-            row[Constant15.base_table_index()] = round_constants[15];
-        }
-        hash_trace_addendum
-    }
-
-    /// The round constants for round `round_number` if it is a valid round number in the Tip5
-    /// permutation, and the zero vector otherwise.
-    fn tip5_round_constants_by_round_number(
-        round_number: usize,
-    ) -> [BFieldElement; NUM_ROUND_CONSTANTS] {
-        match round_number {
-            i if i < tip5::NUM_ROUNDS => tip5::ROUND_CONSTANTS
-                [NUM_ROUND_CONSTANTS * i..NUM_ROUND_CONSTANTS * (i + 1)]
-                .try_into()
-                .unwrap(),
-            _ => [BFIELD_ZERO; NUM_ROUND_CONSTANTS],
-        }
-    }
-
-    /// The inverse-or-zero of (`mid_high` + (`highest` << 16) - (1 << 32) + 1) where `highest`
-    /// is the most significant limb of the given `state_element`, and `mid_high` the second-most
-    /// significant limb.
-    fn inverse_or_zero_of_highest_2_limbs(state_element: BFieldElement) -> BFieldElement {
-        let limbs = state_element.raw_u16s().map(|limb| limb as u64);
-        let highest = limbs[3];
-        let mid_high = limbs[2];
-        let to_invert = mid_high + (highest << 16) - (1 << 32) + 1;
-        BFieldElement::from_raw_u64(to_invert).inverse_or_zero()
     }
 }
